Move submodules from utility to clowder module
