+++ clowder/error/clowder_error.py
