+++ clowder/git/git_submodules.py
@@ -7,9 +7,9 @@ import sys
 from git import GitError
 from termcolor import cprint
 
-from clowder.utility.execute import execute_command
-from clowder.utility.git.git_repo import GitRepo
 import clowder.utility.formatting as fmt
+from clowder.git.git_repo import GitRepo
+from clowder.utility.execute import execute_command
 
 
 class GitSubmodules(GitRepo):
