+++ clowder/git/printing.py
@@ -7,9 +7,9 @@ import sys
 
 from termcolor import colored, cprint
 
-from clowder.utility.execute import execute_command
-from clowder.utility.git.git_repo import GitRepo
 import clowder.utility.formatting as fmt
+from clowder.git.git_repo import GitRepo
+from clowder.utility.execute import execute_command
 
 
 def print_exists(repo_path):
